% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAecomLogo.R
\name{addAecomLogo}
\alias{addAecomLogo}
\title{----------
Add Aecom Logo}
\usage{
addAecomLogo(map, position = "bottomleft")
}
\arguments{
\item{map}{map for the pipe to work}

\item{position}{logo position parameter see addLogo from mapview for options.}
}
\description{
Adds aecom logo to leaflet maps.
}
\keyword{,save,}
\keyword{htmlwidget}
\keyword{map}
