% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_creator_vector.R
\name{convert_hour_to_tp}
\alias{convert_hour_to_tp}
\title{convert_hour_to_tp}
\usage{
convert_hour_to_tp(hour, tp_list = list(AM = 7:9, PM = 16:18, IP =
  10:15))
}
\arguments{
\item{hour}{numeric vector of hours}

\item{tp_list}{a list of tp with the hours starting of the time periods}
}
\description{
Obtains the TP membership from an hour list and a definition of time periods
}
\keyword{am,ip,pm,}
\keyword{hour}
\keyword{tp,}
