% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_functions.R
\name{get_bank_holidaysUK}
\alias{get_bank_holidaysUK}
\title{Get bank holidays for the UK}
\usage{
get_bank_holidaysUK()
}
\value{
a vector of Date Class with bank holidays until from 1998 to 2100
}
\description{
Get bank holidays for the UK
}
