% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emme.R
\name{read_emme}
\alias{read_emme}
\title{read_emme}
\usage{
read_emme(file, rbind_output = "FALSE")
}
\arguments{
\item{file}{The geometry column of an sf object}

\item{rbind_output}{Logical. If true gives the minor direction of the coordinate pair. FALSE by default.}
}
\description{
From a set of geometry it calculates the main direction between the first and last points
aka directionaliser
}
\keyword{emme}
\keyword{rad,}
