% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_to_network.R
\name{robocopy_pfolder}
\alias{robocopy_pfolder}
\title{Robocopy project folder to network}
\usage{
robocopy_pfolder(dest_folder)
}
\arguments{
\item{dest_folder}{destination folder}
}
\description{
Copy project files to another folder and overwrite if the ones you have are newer. Deletes deleted files in the destination folder too. Windows only.
}
\keyword{copy,}
\keyword{sync}
