% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savemaphtml.R
\name{savemaphtml}
\alias{savemaphtml}
\title{save html widget in any folder}
\usage{
savemaphtml(m, path)
}
\arguments{
\item{m}{a map object obtained from leaflet}

\item{path}{path to save the map}
}
\description{
Holla
}
\keyword{,save,}
\keyword{htmlwidget}
\keyword{map}
