% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_to_network.R
\name{sync_to_network}
\alias{sync_to_network}
\title{syncto network}
\usage{
sync_to_network(net_folder)
}
\arguments{
\item{folder}{destination folder}
}
\description{
Copy project files to another folder and overwrite if the ones you have are newer.
}
\keyword{copy,}
\keyword{sync}
