% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_bins.R
\name{default_bins}
\alias{default_bins}
\title{default_bins}
\usage{
default_bins(df, midpoint = 0)
}
\arguments{
\item{df}{a dataframe with only numeric columns or a vector}

\item{midpoint}{0 by efault but can be changed}
}
\description{
It gives you some default bins using jenks breaks and a midpoint of 0
}
\keyword{coordinates,}
\keyword{directions,}
\keyword{geometry,}
\keyword{nb,sb,eb,wb}
