% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_link_direction.R
\name{get_directions}
\alias{get_directions}
\title{get_directions}
\usage{
get_directions(geometry)
}
\description{
From a set of geometry it calculates the main direction between the first and last points
aka directionaliser
 @param geometry The geometry column of an sf object
 @keywords geometry, coordinates, directions, nb,sb,eb,wb
 @export
 @example
 sf_object$direction <- get_directions(sf_object$geometry)
}
