% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emme.R
\name{read_emme}
\alias{read_emme}
\title{read_emme}
\usage{
read_emme(file, rbind_output = "FALSE")
}
\arguments{
\item{file}{file with emme mat format}

\item{rbind_output}{Logical. If true binds the optput into one list.}
}
\description{
Read a emme mat file as a list of the individual matrices
}
\keyword{emme}
\keyword{read,}
