% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tools.R
\name{savemaphtml}
\alias{savemaphtml}
\title{save html widget in any folder}
\usage{
savemaphtml(m, path, ...)
}
\arguments{
\item{m}{a map object obtained from leaflet}

\item{path}{path to save the map}
}
\description{
Wrapper around saveWidgethtml that saves the map in an arbitrary folder using setwd().
}
\keyword{,save,}
\keyword{htmlwidget}
\keyword{map}
